So this is exactly what Claude Code did when given the following prompt.  I haven't tested or read this, but recording it here so that what needs to happen next can be recorded as diffs off this.
The prompt given to Claude Code:

The goal here is to make an excel version of the google apps scripts I copied into the google_version directory.  Add a button to the task pane of the excel add-on that triggers the auto-categorization.  When triggered by the button, our add in shouold look in the workbook for tables  named "Transactions" and "Categories".  If those aren't found, it should stop with an error displayed to the user.  Otherwise, it should proceed to update the “Category” and “Description” columns on all currently visible rows (where “visible” takes into account the currently active filter view) where the “Category” column is blank.  It should use the full set of transactions and categories to make Gemeni API calls in order to get suggested new categories and descriptions.  Look at the the scripts in the google_version directory to see how a google apps script add-on did an equivalent thing in Google sheets.  The same prompt and AI API settings from the Google Apps Script version (in the lookupDescAndCategoryGemeni and lookupDescAndCategoryOpenAI functions) can be used.  My add-on should take the OpenAI or Gemini API keys from dotenv, with the user instructed to put them in their running environment or .env file.